Refactors executeSyncCmd to utils file
